requirements.txt
numpy==1.22.4
Pillow==9.2.0
Pillow==10.3.0
torch==1.11.0
torchvision==0.16.1
tqdm==4.64.0
transformers==4.19.2
wandb==0.17.3
